Split maven package into several files
